+++ src/components/molecules/form/layout.js
@@ -1,0 +1,6 @@
+import { multiply } from '../../_helpers/pixel-calc'
+
+const formWidth = '625px'
+const labelWidth = multiply(formWidth, 0.35)
+
+export { formWidth, labelWidth }
